+++ lib/util/random.h
@@ -18,7 +18,7 @@ extern "C" {
 ///
 /// @param bound Exclusive upper bound on the sequence
 /// @return A permutation of `[0, bound - 1]`
-UTIL_API int *gv_permutation(size_t bound);
+UTIL_API size_t *gv_permutation(size_t bound);
 
 /// generate a random number in the range `[0, bound - 1]`
 ///
